:recycle: refactor: useApp custom hook으로 분리
